rewrote some parts of the (almost deprecated) textmeshgenerator, created tool for packaging fonts in a way that makes it easy and fast to read for the engine, rewrote parts of Font , fixed some alignment stuff

+++ Walgelijk/Text/Kerning.cs
@@ -16,6 +16,6 @@
         /// <summary>
         /// Offset amount
         /// </summary>
-        public int Amount;
+        public float Amount;
     }
 }
